Refactor: App 내 관리자, 일반 페이지 분리.
