+++ src/basic/components/icons/index.tsx
@@ -1,0 +1,12 @@
+// TODO: SVG 아이콘 컴포넌트들
+// 구현할 아이콘:
+// - CartIcon: 장바구니 아이콘
+// - AdminIcon: 관리자 아이콘
+// - PlusIcon: 플러스 아이콘
+// - MinusIcon: 마이너스 아이콘
+// - TrashIcon: 삭제 아이콘
+// - ChevronDownIcon: 아래 화살표
+// - ChevronUpIcon: 위 화살표
+// - CheckIcon: 체크 아이콘
+
+// TODO: 구현
