+++ src/basic/constants/index.ts
@@ -1,0 +1,8 @@
+// TODO: 초기 데이터 상수
+// 정의할 상수들:
+// - initialProducts: 초기 상품 목록 (상품1, 상품2, 상품3 + 설명 필드 포함)
+// - initialCoupons: 초기 쿠폰 목록 (5000원 할인, 10% 할인)
+//
+// 참고: origin/App.tsx의 초기 데이터 구조를 참조
+
+// TODO: 구현
