+++ src/basic/hooks/useCart.ts
@@ -1,0 +1,29 @@
+// TODO: 장바구니 관리 Hook
+// 힌트:
+// 1. 장바구니 상태 관리 (localStorage 연동)
+// 2. 상품 추가/삭제/수량 변경
+// 3. 쿠폰 적용
+// 4. 총액 계산
+// 5. 재고 확인
+//
+// 사용할 모델 함수:
+// - cartModel.addItemToCart
+// - cartModel.removeItemFromCart
+// - cartModel.updateCartItemQuantity
+// - cartModel.calculateCartTotal
+// - cartModel.getRemainingStock
+//
+// 반환할 값:
+// - cart: 장바구니 아이템 배열
+// - selectedCoupon: 선택된 쿠폰
+// - addToCart: 상품 추가 함수
+// - removeFromCart: 상품 제거 함수
+// - updateQuantity: 수량 변경 함수
+// - applyCoupon: 쿠폰 적용 함수
+// - calculateTotal: 총액 계산 함수
+// - getRemainingStock: 재고 확인 함수
+// - clearCart: 장바구니 비우기 함수
+
+export function useCart() {
+  // TODO: 구현
+}
