+++ src/basic/utils/hooks/useDebounce.ts
@@ -1,0 +1,11 @@
+// TODO: 디바운스 Hook
+// 힌트:
+// 1. 값이 변경되어도 지정된 시간 동안 대기
+// 2. 대기 시간 동안 값이 다시 변경되면 타이머 리셋
+// 3. 최종적으로 안정된 값만 반환
+//
+// 사용 예시: 검색어 입력 디바운싱
+
+export function useDebounce<T>(value: T, delay: number): T {
+  // TODO: 구현
+}
