+++ src/basic/utils/hooks/useLocalStorage.ts
@@ -1,0 +1,15 @@
+// TODO: LocalStorage Hook
+// 힌트:
+// 1. localStorage와 React state 동기화
+// 2. 초기값 로드 시 에러 처리
+// 3. 저장 시 JSON 직렬화/역직렬화
+// 4. 빈 배열이나 undefined는 삭제
+//
+// 반환값: [저장된 값, 값 설정 함수]
+
+export function useLocalStorage<T>(
+  key: string,
+  initialValue: T
+): [T, (value: T | ((val: T) => T)) => void] {
+  // TODO: 구현
+}
