+++ src/basic_original/__tests__/origin.test.tsx
@@ -1,7 +1,6 @@
 // @ts-nocheck
 import { render, screen, fireEvent, within, waitFor } from '@testing-library/react';
 import { vi } from 'vitest';
-
 import App from '../App';
 import '../../setupTests';
