+++ src/basic_original/main.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-
 import App from './App.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
