+++ src/origin/main.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+
 import App from './App.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
